My first commit

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -27,6 +27,6 @@ size_t print_dlistint(const dlistint_t *h)
         h = h->next;
     }
 
-    return node_count;
+    return (size_t)node_count;
 }
